Moved utils to a new folder
